requirements.txt
python-telegram-bot==20.2
python-dotenv==1.0.0
psycopg==3.1.8